.claude/worktrees/keen-shaw-651613/fastapi-backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0
greenlet>=3.0.0                  # required by SQLAlchemy async on Python 3.12+
alembic==1.13.1
psycopg2-binary==2.9.9          # sync driver (Alembic / scripts)

# ── Redis ──────────────────────────────────────────────────────────────────────
redis[hiredis]==5.0.4

# ── Auth ───────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.1.2                    # used directly (passlib incompatible with Python 3.12+)

# ── Validation & Settings ──────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ── HTTP Client (outbound calls to AI service) ────────────────────────────────
httpx==0.27.0

# ── AI / LangChain ─────────────────────────────────────────────────────────────
langchain==0.1.20
langchain-openai==0.1.7
openai==1.30.1

# ── Rate limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9
limits==3.10.1

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
orjson==3.10.3
pytz==2024.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0                    # also used for TestClient
